Refactor imports and add data transformation functions for eidolon value analysis
